Add startswith filter and move entityrelation to new package entity.relation
