+++ dc-model-feign/pom.xml
@@ -5,7 +5,7 @@
   <parent>
     <groupId>de.digitalcollections.model</groupId>
     <artifactId>dc-model-parent</artifactId>
-    <version>8.1.1-SNAPSHOT</version>
+    <version>8.2.0-SNAPSHOT</version>
   </parent>
 
   <name>DigitalCollections: Model (Feign)</name>
