+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/DigitalCollectionsModelModule.java
@@ -16,7 +16,6 @@ import de.digitalcollections.model.api.identifiable.entity.Article;
 import de.digitalcollections.model.api.identifiable.entity.Collection;
 import de.digitalcollections.model.api.identifiable.entity.DigitalObject;
 import de.digitalcollections.model.api.identifiable.entity.Entity;
-import de.digitalcollections.model.api.identifiable.entity.EntityRelation;
 import de.digitalcollections.model.api.identifiable.entity.Project;
 import de.digitalcollections.model.api.identifiable.entity.Topic;
 import de.digitalcollections.model.api.identifiable.entity.Website;
@@ -29,6 +28,8 @@ import de.digitalcollections.model.api.identifiable.entity.geo.GeoLocation;
 import de.digitalcollections.model.api.identifiable.entity.geo.HumanSettlement;
 import de.digitalcollections.model.api.identifiable.entity.parts.Subtopic;
 import de.digitalcollections.model.api.identifiable.entity.parts.Webpage;
+import de.digitalcollections.model.api.identifiable.entity.relation.EntityRelation;
+import de.digitalcollections.model.api.identifiable.entity.relation.Predicate;
 import de.digitalcollections.model.api.identifiable.entity.work.Expression;
 import de.digitalcollections.model.api.identifiable.entity.work.Item;
 import de.digitalcollections.model.api.identifiable.entity.work.Manifestation;
@@ -70,7 +71,6 @@ import de.digitalcollections.model.api.paging.PageResponse;
 import de.digitalcollections.model.api.paging.SearchPageRequest;
 import de.digitalcollections.model.api.paging.SearchPageResponse;
 import de.digitalcollections.model.api.paging.Sorting;
-import de.digitalcollections.model.api.relations.Predicate;
 import de.digitalcollections.model.api.security.User;
 import de.digitalcollections.model.api.view.BreadcrumbNavigation;
 import de.digitalcollections.model.impl.view.RenderingHints;
@@ -86,7 +86,6 @@ import de.digitalcollections.model.jackson.mixin.identifiable.entity.ArticleMixI
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.CollectionMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.DigitalObjectMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityMixIn;
-import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityRelationMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.ProjectMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.TopicMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.WebsiteMixIn;
@@ -99,6 +98,8 @@ import de.digitalcollections.model.jackson.mixin.identifiable.entity.geo.GeoLoca
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.geo.HumanSettlementMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.parts.SubtopicMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.parts.WebpageMixIn;
+import de.digitalcollections.model.jackson.mixin.identifiable.entity.relation.EntityRelationMixIn;
+import de.digitalcollections.model.jackson.mixin.identifiable.entity.relation.PredicateMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.work.ExpressionMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.work.ItemMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.work.ManifestationMixIn;
@@ -139,7 +140,6 @@ import de.digitalcollections.model.jackson.mixin.paging.PageResponseMixIn;
 import de.digitalcollections.model.jackson.mixin.paging.SearchPageRequestMixIn;
 import de.digitalcollections.model.jackson.mixin.paging.SearchPageResponseMixIn;
 import de.digitalcollections.model.jackson.mixin.paging.SortingMixIn;
-import de.digitalcollections.model.jackson.mixin.relations.PredicateMixIn;
 import de.digitalcollections.model.jackson.mixin.security.UserMixIn;
 import de.digitalcollections.model.jackson.mixin.view.BreadcrumbNavigationMixIn;
 import de.digitalcollections.model.jackson.mixin.view.RenderingHintsMixIn;
