+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/entity/relation/PredicateMixIn.java
@@ -1,8 +1,8 @@
-package de.digitalcollections.model.jackson.mixin.relations;
+package de.digitalcollections.model.jackson.mixin.identifiable.entity.relation;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.impl.relations.PredicateImpl;
+import de.digitalcollections.model.impl.identifiable.entity.relation.PredicateImpl;
 
 @JsonDeserialize(as = PredicateImpl.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
