+++ dc-model/pom.xml
@@ -6,7 +6,7 @@
   <parent>
     <groupId>de.digitalcollections.model</groupId>
     <artifactId>dc-model-parent</artifactId>
-    <version>8.1.1-SNAPSHOT</version>
+    <version>8.2.0-SNAPSHOT</version>
   </parent>
 
   <name>DigitalCollections: Model (Model)</name>
